fix(pages):making the code more readable
