+++ src/home.js
@@ -1,0 +1,11 @@
+const home = ()=>{
+    let lorem = document.createElement('p');
+    lorem.innerText = `Lorem ipsum dolor sit amet, consectetur adipiscing elit. Aliquam et congue velit. Aenean nec viverra elit. Donec vitae mi pulvinar, lacinia nulla sed, eleifend felis. In sagittis tortor non dui venenatis eleifend. Cras placerat, libero nec dignissim euismod, est mi convallis erat, non congue justo leo eget dui. Proin congue quam vitae tristique fringilla. Donec sit amet cursus sapien. Curabitur ut dictum orci. Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere cubilia Curae; Ut vitae quam dui. Integer ac mi tincidunt elit lobortis varius. Ut varius iaculis mi sit amet posuere. Proin sit amet sem lacus. Integer posuere, magna vel convallis laoreet, nunc velit volutpat enim, ut lacinia nunc libero vitae velit. Etiam aliquet libero sit amet purus luctus egestas.
+
+    Fusce venenatis nisi vitae erat mollis, a facilisis orci vulputate. Nam lectus felis, ornare consectetur nisi et, convallis bibendum leo. Vestibulum et posuere enim. Phasellus facilisis eros arcu, vitae suscipit magna semper nec. Donec ornare placerat mattis. Proin eget ultricies risus. Nam libero mi, venenatis commodo vehicula euismod, tristique ac odio. Maecenas felis lorem, tempor fringilla feugiat eu, faucibus et nunc. Vestibulum dignissim sem quis lectus rutrum bibendum. Praesent turpis ex, feugiat vitae aliquam hendrerit, viverra at tellus. Quisque lacinia dictum quam nec pretium.
+    
+    Vestibulum quis orci et velit viverra malesuada in at justo. Integer pharetra tempor mauris in mattis. Vivamus ornare magna quis turpis varius hendrerit. Vestibulum dui magna, gravida eget iaculis ultricies, congue sed mi. Praesent dictum euismod metus, ac feugiat mauris. Aliquam pharetra est ipsum, in semper magna porttitor nec. Nulla ac magna vel lorem aliquam suscipit. Donec quis dolor eu ante sollicitudin fermentum. Fusce varius vulputate justo, non ornare ligula scelerisque nec.`
+    return lorem;
+}
+
+export{home}
